extract create address book helper method
